6-size.c

+++ 0x00-hello_world/6-size.c
@@ -11,7 +11,7 @@ int main(void)
 	printf("Size of a char:%lu bytes\n", sizeof(char));
 	printf("Size of an int: %lu bytes\n", sizeof(int));
 	printf("Size of a long int: %lu bytes\n", sizeof(long));
-	printf("Size of a long long int:%lu bytes\n", sizeof(long));
+	printf("Size of a long long int:%lu bytes\n", sizeof(long long));
 	printf("Size of a float:%lu bytes\n", sizeof(float));
 
 	return (0);
